# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admins": {
      "post": {
        "operationId": "AdminsController_createAdmin",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Admin user created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Admin user created successfully",
                    "user": {
                      "id": 1,
                      "email": "admin@example.com",
                      "username": "admin1",
                      "name": "Admin Name",
                      "role": "ADMIN"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Admin with this email or username already exists"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions (must be SUPER_ADMIN)"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a new admin user",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admins/{id}": {
      "delete": {
        "operationId": "AdminsController_deleteAdmin",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Admin user ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin user deleted successfully"
          },
          "400": {
            "description": "Bad Request - Admin not found or attempting to delete a Super Admin"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions or trying to delete another admin"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete an admin user",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admins/students/{id}/approve": {
      "patch": {
        "operationId": "AdminsController_approveStudent",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Student approved successfully"
          },
          "400": {
            "description": "Bad Request - Student not found, not a student, or already approved"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Approve a student account",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admins/students/{id}/reject": {
      "patch": {
        "operationId": "AdminsController_rejectStudent",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Student rejected successfully"
          },
          "400": {
            "description": "Bad Request - Student not found, not a student, already rejected, or already approved"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Reject a student account",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admins/categories": {
      "post": {
        "operationId": "AdminsController_createCategory",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": 1,
                    "name": "Category Name",
                    "description": "Category Description",
                    "createdAt": "2024-04-09T12:00:00Z",
                    "updatedAt": "2024-04-09T12:00:00Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input data"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "409": {
            "description": "Conflict - Category with this name already exists"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a new category",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admins/categories/{id}": {
      "delete": {
        "operationId": "AdminsController_deleteCategory",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted successfully"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Not Found - Category not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a category",
        "tags": [
          "Admin"
        ]
      },
      "patch": {
        "operationId": "AdminsController_updateCategory",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": 1,
                    "name": "Updated Category Name",
                    "description": "Updated Category Description",
                    "updatedAt": "2024-04-09T12:00:00Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input data"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Not Found - Category not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a category",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admins/users/students": {
      "get": {
        "operationId": "AdminsController_getAllStudents",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by approval status",
            "schema": {
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of students retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": [
                      {
                        "id": 1,
                        "email": "student@example.com",
                        "username": "student1",
                        "name": "Student Name",
                        "approvalStatus": "PENDING"
                      }
                    ],
                    "meta": {
                      "total": 1,
                      "page": 1,
                      "limit": 10
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get all students with optional filtering",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admins/users/students/search": {
      "get": {
        "description": "Search by name or academic number",
        "operationId": "AdminsController_searchStudents",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Bearer token for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": true,
            "in": "query",
            "description": "Search term",
            "schema": {
              "example": "123",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Student search results",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": 4,
                      "name": "Name",
                      "email": "example@gmail.com",
                      "student": {
                        "id": 3,
                        "userId": 4,
                        "academicNumber": "1112345678910",
                        "department": "IT",
                        "studyLevel": 1,
                        "approvalStatus": "PENDING",
                        "approvalUpdatedByUserId": null
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": [
                      "query must be a string",
                      "page must not be less than 1",
                      "limit must not be greater than 50"
                    ],
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Admin access required"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Search students",
        "tags": [
          "Admin"
        ]
      }
    }
  },
  "info": {
    "title": "SAHIM API Documentation",
    "description": "SAHIM API is the backend service that powers the SAHIM platform.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken",
        "description": "Refresh token for authentication"
      }
    },
    "schemas": {
...
      "AdminSignupDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@example.com",
            "maxLength": 255
          },
          "username": {
            "type": "string",
            "description": "Username (letters, numbers, underscores and hyphens only)",
            "example": "johndoe",
            "minLength": 3,
            "maxLength": 50
          },
          "name": {
            "type": "string",
            "description": "User full name",
            "example": "John Doe",
            "minLength": 3,
            "maxLength": 100
          },
          "password": {
            "type": "string",
            "description": "Password (must contain uppercase, lowercase, number and special character)",
            "example": "Password123!",
            "minLength": 8,
            "maxLength": 72
          },
          "role": {
            "type": "string",
            "description": "Admin role (defaults to ADMIN)",
            "enum": [
              "ADMIN",
              "STUDENT",
              "SUPER_ADMIN"
            ],
            "default": "ADMIN"
          }
        },
        "required": [
          "email",
          "username",
          "name",
          "password"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Category name",
            "example": "Programming",
            "minLength": 2,
            "maxLength": 50
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Updated category name",
            "example": "Web Development",
            "minLength": 2,
            "maxLength": 50
          }
        },
        "required": [
          "name"
        ]
      }
    }
  }
}